Move tests from vehicle_status to vehicle
