+++ bimmer_connected/tests/test_vehicle.py
@@ -5,6 +5,7 @@ import respx
 from bimmer_connected.const import ATTR_ATTRIBUTES, ATTR_STATE, CarBrands
 from bimmer_connected.models import GPSPosition, StrEnum, VehicleDataBase
 from bimmer_connected.vehicle import VehicleViewDirection
+from bimmer_connected.vehicle.charging_sessions import ChargingType
 from bimmer_connected.vehicle.const import DriveTrainType
 from bimmer_connected.vehicle.reports import CheckControlMessageReport
 
@@ -285,3 +286,53 @@ def test_gpsposition():
     assert pos == (1, 2)
     assert pos != "(1, 2)"
     assert pos[0] == 1
+
+
+@pytest.mark.asyncio
+async def test_charging_statistics(caplog, bmw_fixture: respx.Router):
+    """Test if the parsing of charging statistics is working."""
+
+    status = (await prepare_account_with_vehicles()).get_vehicle(VIN_U11).charging_statistics
+
+    if not status:
+        pytest.skip("No charging statistics available")
+
+    assert status.charging_session_timeperiod == "September 2023"
+    assert status.charging_session_count == 6
+    assert status.total_energy_charged == 168
+
+    assert len(get_deprecation_warning_count(caplog)) == 0
+
+
+@pytest.mark.asyncio
+async def test_charging_sessions(caplog, bmw_fixture: respx.Router):
+    """Test if the parsing of charging sessions is working."""
+
+    status = (await prepare_account_with_vehicles()).get_vehicle(VIN_U11).charging_sessions
+
+    if not status:
+        pytest.skip("No charging sessions available")
+
+    assert status.charging_session_count == "6"
+    assert status.charging_sessions[0]["status"] == "FINISHED"
+    assert status.charging_sessions[0]["description"] == "9/3/2023 15:47 • some_road • duration • -- EUR"
+    assert status.charging_sessions[0]["address"] == "N/A"
+    assert status.charging_sessions[0]["charging_type"] == "N/A"
+    assert status.charging_sessions[0]["soc_start"] == "N/A"
+    assert status.charging_sessions[0]["soc_end"] == "N/A"
+    assert status.charging_sessions[0]["energy_charged"] == "N/A"
+    assert status.charging_sessions[0]["date_start"] == "N/A"
+    assert status.charging_sessions[0]["date_end"] == "N/A"
+    assert status.charging_sessions[0]["duration"] == "N/A"
+    assert status.charging_sessions[0]["charging_speed_min"] == "N/A"
+    assert status.charging_sessions[0]["charging_speed_max"] == "N/A"
+    assert status.charging_sessions[0]["public"] is True
+    assert status.charging_sessions[0]["pre_condition"] == "N/A"
+    assert status.charging_sessions[0]["mileage"] == "N/A"
+
+    assert len(get_deprecation_warning_count(caplog)) == 0
+
+
+def test_charging_type_unknown():
+    unknown_charging_type = ChargingType("ABCs")
+    assert unknown_charging_type == ChargingType.UNKNOWN
